refactor: add useFilter, simplify code structure
