+++ components/Table/hooks/useSorter.tsx
@@ -1,10 +1,89 @@
 import { useState, useRef, useCallback } from 'react';
-import { SorterInfo } from '../interface';
-import { getSorterPriority, getSorterFn } from '../utils';
+import { SorterInfo, SorterFn, SortDirection, InternalColumnProps } from '../interface';
+import { getSorterPriority, getSorterFn, getColumnByUniqueKey } from '../utils';
 import useUpdate from '../../_util/hooks/useUpdate';
-import { isNumber } from '../../_util/is';
+import { isNumber, isArray } from '../../_util/is';
 
-export default function useSorter(flattenColumns, defaultSorters) {
+function getDefaultSorters(flattenColumns: InternalColumnProps[]) {
+  let defaultSorters: SorterInfo[] = [];
+  flattenColumns.forEach((column) => {
+    const innerDataIndex = column.key;
+
+    if ('defaultSortOrder' in column || 'sortOrder' in column) {
+      const priority = getSorterPriority(column.sorter);
+      const direction = 'sortOrder' in column ? column.sortOrder : column.defaultSortOrder;
+      const sorter: SorterInfo = {
+        field: innerDataIndex,
+        direction,
+        sorterFn: getSorterFn(column.sorter),
+        priority,
+      };
+      if (!direction) {
+        defaultSorters.push(sorter);
+      } else if (isNumber(priority)) {
+        if (defaultSorters.every((item) => isNumber(item.priority) || !item.direction)) {
+          defaultSorters.push(sorter);
+        }
+      } else if (defaultSorters.every((item) => !item.direction)) {
+        defaultSorters.push(sorter);
+      } else {
+        defaultSorters = [sorter];
+      }
+    }
+  });
+
+  return defaultSorters;
+}
+
+function compareFn(activeSorters: SorterInfo[]) {
+  const compare = function (fn: SorterFn, direction: SortDirection) {
+    return (a, b) => {
+      const result = fn(a, b);
+      return direction === 'descend' ? -result : result;
+    };
+  };
+  const sorters = [...activeSorters];
+  sorters.sort((a, b) => b.priority - a.priority);
+  return (a, b) => {
+    for (let i = 0, l = sorters.length; i < l; i++) {
+      const { sorterFn, direction } = sorters[i];
+      if (typeof sorterFn !== 'function') {
+        continue;
+      }
+      const result = compare(sorterFn, direction)(a, b);
+      if (result !== 0) {
+        return result;
+      }
+    }
+    return 0;
+  };
+}
+
+const getSortedDataImpl = (data, activeSorters: SorterInfo[], childrenColumnName: string) => {
+  return data
+    .slice()
+    .sort(compareFn(activeSorters))
+    .map((item) => {
+      if (isArray(item[childrenColumnName])) {
+        return {
+          ...item,
+          [childrenColumnName]: getSortedDataImpl(
+            item[childrenColumnName],
+            activeSorters,
+            childrenColumnName
+          ),
+        };
+      }
+      return item;
+    });
+};
+
+export default function useSorter(
+  flattenColumns: InternalColumnProps[],
+  childrenColumnName: string,
+  handleOnChange
+) {
+  const defaultSorters = getDefaultSorters(flattenColumns);
   const [activeSorters, setActiveSorters] = useState<SorterInfo[]>(defaultSorters);
   const [currentSorter, setCurrentSorter] = useState<SorterInfo>({});
   const prevFlattenColumnsRef = useRef(flattenColumns);
@@ -29,6 +108,16 @@ export default function useSorter(flattenColumns, defaultSorters) {
     },
     [activeSorters]
   );
+
+  const getSortedData = (data) => {
+    if (
+      (currentSorter.direction && typeof currentSorter.sorterFn === 'function') ||
+      activeSorters.length
+    ) {
+      return getSortedDataImpl(data, activeSorters, childrenColumnName);
+    }
+    return data;
+  };
 
   const getControlledSorters = useCallback((columns) => {
     const controlledColumns = columns.filter((column) => 'sortOrder' in column);
@@ -68,6 +157,24 @@ export default function useSorter(flattenColumns, defaultSorters) {
     [flattenColumns, getControlledSorters, setActiveSorters, setCurrentSorter]
   );
 
+  function onSort(direction: SortDirection, field) {
+    const column = getColumnByUniqueKey(flattenColumns, field);
+    if (!column) {
+      return;
+    }
+    const sorter: SorterInfo = {
+      direction,
+      field,
+      sorterFn: getSorterFn(column.sorter),
+      priority: getSorterPriority(column.sorter),
+    };
+    const nextActiveSorters = getNextActiveSorters(sorter);
+    updateStateSorters(sorter, nextActiveSorters);
+    handleOnChange('sort', {
+      sorters: [sorter],
+    });
+  }
+
   useUpdate(() => {
     const prevFlattenColumns = prevFlattenColumnsRef.current;
     const prevControlledSorters = getControlledSorters(prevFlattenColumns);
@@ -101,7 +208,7 @@ export default function useSorter(flattenColumns, defaultSorters) {
   return {
     currentSorter,
     activeSorters,
-    getNextActiveSorters,
-    updateStateSorters,
+    onSort,
+    getSortedData,
   };
 }
