+++ components/Table/interface.tsx
@@ -605,7 +605,6 @@ export interface TheadProps<T = any> {
   data: T[];
   onHandleFilter: (column, filter) => void;
   onHandleFilterReset: (filter) => void;
-  currentSorter: SorterInfo;
   activeSorters: SorterInfo[];
   selectedRowKeys: (string | number)[];
   onHeaderRow?: (columns, index: number) => RowCallbackProps;
@@ -686,3 +685,5 @@ export interface SummaryProps {
   fixed?: 'top' | 'bottom';
   children?: ReactNode;
 }
+
+export type FilterType<T> = Partial<Record<keyof T, string[]>>;
